added MtdWriter and used it in asahi-nvram utility
